--- a/agent/agent.go
+++ b/agent/agent.go
@@ -957,6 +957,7 @@ func (a *Agent) consulConfig() (*consul.Config, error) {
 		base.SerfWANConfig = nil
 	}
 
+	base.EnableGRPC = a.config.ServerEnableGRPC
 	base.RPCAddr = a.config.RPCBindAddr
 	base.RPCAdvertise = a.config.RPCAdvertiseAddr