--- a/agent/config/builder.go
+++ b/agent/config/builder.go
@@ -837,6 +837,7 @@ func (b *Builder) Build() (rt RuntimeConfig, err error) {
 		PrimaryDatacenter:                       primaryDatacenter,
 		RPCAdvertiseAddr:                        rpcAdvertiseAddr,
 		RPCBindAddr:                             rpcBindAddr,
+		ServerEnableGRPC:                        b.boolValWithDefault(c.ServerEnableGRPC, true),
 		RPCHoldTimeout:                          b.durationVal("performance.rpc_hold_timeout", c.Performance.RPCHoldTimeout),
 		RPCMaxBurst:                             b.intVal(c.Limits.RPCMaxBurst),
 		RPCProtocol:                             b.intVal(c.RPCProtocol),