--- a/agent/config/runtime.go
+++ b/agent/config/runtime.go
@@ -894,6 +894,11 @@ type RuntimeConfig struct {
 	// hcl: bind_addr = string ports { server = int }
 	RPCBindAddr *net.TCPAddr
 
+	// ServerEnableGRPC controls wether to enable GRPC for RPC communications
+	//
+	// hcl: bind_addr = string ports { sever_grpc = int }
+	ServerEnableGRPC bool
+
 	// RPCHoldTimeout is how long an RPC can be "held" before it is errored.
 	// This is used to paper over a loss of leadership by instead holding RPCs,
 	// so that the caller experiences a slow response rather than an error.