--- a/agent/consul/health_endpoint.go
+++ b/agent/consul/health_endpoint.go
@@ -152,7 +152,7 @@ func (h *Health) ServiceNodes(args *structs.ServiceSpecificRequest, reply *struc
 	}
 
 	// Determine the function we'll call
-	var f func(memdb.WatchSet, *state.Store, *structs.ServiceSpecificRequest) (uint64, structs.CheckServiceNodes, error)
+	var f func(memdb.WatchSet, *state.Store, *structs.ServiceSpecificRequest) (uint64, []structs.CheckServiceNode, error)
 	switch {
 	case args.Connect:
 		f = h.serviceNodesConnect
@@ -204,7 +204,7 @@ func (h *Health) ServiceNodes(args *structs.ServiceSpecificRequest, reply *struc
 			if err != nil {
 				return err
 			}
-			reply.Nodes = raw.(structs.CheckServiceNodes)
+			reply.Nodes = raw.([]structs.CheckServiceNode)
 
 			return h.srv.sortNodesByDistanceFrom(args.Source, reply.Nodes)
 		})
@@ -247,11 +247,11 @@ func (h *Health) ServiceNodes(args *structs.ServiceSpecificRequest, reply *struc
 // The serviceNodes* functions below are the various lookup methods that
 // can be used by the ServiceNodes endpoint.
 
-func (h *Health) serviceNodesConnect(ws memdb.WatchSet, s *state.Store, args *structs.ServiceSpecificRequest) (uint64, structs.CheckServiceNodes, error) {
+func (h *Health) serviceNodesConnect(ws memdb.WatchSet, s *state.Store, args *structs.ServiceSpecificRequest) (uint64, []structs.CheckServiceNode, error) {
 	return s.CheckConnectServiceNodes(ws, args.ServiceName)
 }
 
-func (h *Health) serviceNodesTagFilter(ws memdb.WatchSet, s *state.Store, args *structs.ServiceSpecificRequest) (uint64, structs.CheckServiceNodes, error) {
+func (h *Health) serviceNodesTagFilter(ws memdb.WatchSet, s *state.Store, args *structs.ServiceSpecificRequest) (uint64, []structs.CheckServiceNode, error) {
 	// DEPRECATED (singular-service-tag) - remove this when backwards RPC compat
 	// with 1.2.x is not required.
 	// Agents < v1.3.0 populate the ServiceTag field. In this case,
@@ -262,6 +262,6 @@ func (h *Health) serviceNodesTagFilter(ws memdb.WatchSet, s *state.Store, args *
 	return s.CheckServiceTagNodes(ws, args.ServiceName, args.ServiceTags)
 }
 
-func (h *Health) serviceNodesDefault(ws memdb.WatchSet, s *state.Store, args *structs.ServiceSpecificRequest) (uint64, structs.CheckServiceNodes, error) {
+func (h *Health) serviceNodesDefault(ws memdb.WatchSet, s *state.Store, args *structs.ServiceSpecificRequest) (uint64, []structs.CheckServiceNode, error) {
 	return s.CheckServiceNodes(ws, args.ServiceName)
 }