--- a/agent/consul/server_oss.go
+++ b/agent/consul/server_oss.go
@@ -1,5 +1,7 @@
 package consul
 
+import "google.golang.org/grpc"
+
 func init() {
 	registerEndpoint(func(s *Server) interface{} { return &ACL{s} })
 	registerEndpoint(func(s *Server) interface{} { return &Catalog{s} })
@@ -7,6 +9,9 @@ func init() {
 	registerEndpoint(func(s *Server) interface{} { return &ConfigEntry{s} })
 	registerEndpoint(func(s *Server) interface{} { return &ConnectCA{srv: s} })
 	registerEndpoint(func(s *Server) interface{} { return &Health{s} })
+	registerGRPCEndpoint(func(s *Server, grpcServer *grpc.Server) {
+		RegisterHealthServer(grpcServer, &HealthGRPCAdapter{Health{s}})
+	})
 	registerEndpoint(func(s *Server) interface{} { return &Intention{s} })
 	registerEndpoint(func(s *Server) interface{} { return &Internal{s} })
 	registerEndpoint(func(s *Server) interface{} { return &KVS{s} })