… agent/config/runtime_test.go
… agent/consul/acl_test.go
… agent/consul/client_test.go
… agent/consul/health_endpoint.pb.go
… agent/consul/health_endpoint_test.go
… agent/consul/internal_endpoint_test.go
… agent/consul/prepared_query_endpoint_test.go
… agent/consul/rtt_test.go
… agent/consul/server.go
… agent/consul/server_serf.go
… agent/consul/server_test.go
… agent/consul/snapshot_endpoint.go
… agent/consul/snapshot_endpoint_test.go
… agent/consul/state/catalog.go
… agent/consul/stats_fetcher.go
… agent/dns.go
… agent/health_endpoint.go
… agent/health_endpoint_test.go
… agent/metadata/server.go
… agent/pool/conn.go
… agent/prepared_query_endpoint.go
… agent/prepared_query_endpoint_test.go
… agent/proxycfg/manager_test.go
… agent/proxycfg/snapshot.go
… agent/proxycfg/state.go
… agent/proxycfg/testing.go
… agent/structs/check_definition_test.go
… agent/structs/check_type.go
… agent/structs/connect_proxy_config.go
… agent/structs/connect_proxy_config_test.go