This PR aims a starting replacing the current RPC protocol with GRPC. It starts by implementing the `Health.ServiceNodes` call as a POC.

Rationale: 
* Protobuf provides better performance and and smaller serialization size. From my very early (more on that later) benchmarks on a Health.ServiceNode result with ~1K entries, serialization is arround 4x faster and 2x smaller.
* The current RPC library is no longer maintained
* GRPC has built-in support for connection pooling and pipelining via HTTP/2

### Structs conversion
The structs `ServiceSpecificRequet` and `IndexedCheckServiceNodes` as well as the structs used by thoose two have been converted to protobuf. Gogoprotos extentions were used to produce the same Go strucsts than the original to ensure compat with the raft state, older clients etc. 
Two types cound not be mapped to protobuf, even with gogo extentions : `map[string]interface{}` and `map[string][]string`. Custom proto types have been implemented for thoose.
The only exception are the custom slice types such as `type CheckServiceNodes []CheckServiceNode` that are now plain slices, but that does not cause any issues as they marshal / unmarshal the same way.

### GRPC Server
The servers advertise that they support GRPC by providing their `grpc_port` (8310 by default) in Serf. When making a call the agent first checks that the server supports GRPC, and that the method is implemented (currently on `Health.ServiceNodes` is, and the implemented list is hardcoded on the client side).
If both conditions are true then GRPC is used, classic RPC is used.
The behavior is the same for server to server and cross DC calls.   

### TODO
- [ ] More benchmarking with various use cases
- [ ] TLS
- [ ] Retry on failure on par with RPC
- [ ] Cleanup leftover custom slice types
- [ ] Metrics