@pierresouchay 2019-04-12
@banks here is the POC I discussed with you. We plan to test it in real world (just for health to see real world impact in our largest DC) if you would consider it for inclusion eventually