@Aestek 2019-04-15
I'added some benchmarks, here are the results :
```
benchmark                                            old ns/op     new ns/op     delta
BenchmarkIndexedCheckserviceNodesEncoding/10-8       175084        79853         -54.39%
BenchmarkIndexedCheckserviceNodesEncoding/100-8      1718008       768297        -55.28%
BenchmarkIndexedCheckserviceNodesEncoding/500-8      8649143       3834982       -55.66%
BenchmarkIndexedCheckserviceNodesEncoding/1000-8     17097160      7758730       -54.62%

benchmark                                            old allocs     new allocs     delta
BenchmarkIndexedCheckserviceNodesEncoding/10-8       3337           547            -83.61%
BenchmarkIndexedCheckserviceNodesEncoding/100-8      33217          5407           -83.72%
BenchmarkIndexedCheckserviceNodesEncoding/500-8      166017         27009          -83.73%
BenchmarkIndexedCheckserviceNodesEncoding/1000-8     332017         54013          -83.73%

benchmark                                            old bytes     new bytes     delta
BenchmarkIndexedCheckserviceNodesEncoding/10-8       57385         28714         -49.96%
BenchmarkIndexedCheckserviceNodesEncoding/100-8      569852        281963        -50.52%
BenchmarkIndexedCheckserviceNodesEncoding/500-8      2847478       1399372       -50.86%
BenchmarkIndexedCheckserviceNodesEncoding/1000-8     5694503       2790210       -51.00%
```