@Aestek 2019-04-15
## Load testing
We ran some tests on preproduction, here are the results : 

We deployed the version with GRPC on one server and one client, then ran apache benchmark (concurrency 800, forever) against `/v1/health/service/consul-agent-http?stale`. This benchmark highlights the encoding differences as the queries are not blocking.
As the client switched server over time we observed the difference in load and rps (sorry for the ugly prometheus graphs) : 

Load : 
![image](https://user-images.githubusercontent.com/1712219/56147546-cbcf0f80-5fa8-11e9-8850-a8932ecc3851.png)
We can see a significant load reduction on consu03 of ~40%

RPS:
![image](https://user-images.githubusercontent.com/1712219/56147584-dd181c00-5fa8-11e9-9bb2-277131ec35c5.png)
We can see a ~78% RPS increase 

Response time: 
This is by no mean scientific as we have no metrics for this bu by looking at the logs the response time seemed much lower (~800ms vs 1.1s). 

## Analysis
As stated before this benchmark focuses on encoding performance, however we are confident that this will result in performance improvements for Consul in the real world. 
The server running this patch ran fine along with other server and no disruption of service was experienced.