@banks 2019-04-23
@pierresouchay @Aestek, thanks for this work, this is really useful data. We will consider this along with our roadmap of other important performance work.

If we incorporated gRPC progressively for real I think we'd want to use the same RPC port we have now rather than need a whole new one to be opened - we already switch between 3 different protocols over that port based on the first byte. Last time I looked though there were some intricacies with getting Go's gRPC server lib to work with a connection established by an external listener so we'd have to look into that.