@Aestek 2019-04-26
@banks GRPC now uses the same port as classic RPC. Though it would be hacky but it was pretty straightforward.