@Aestek 2019-06-11
## Production results
We deployed GRPC on most of our agents and one server in our biggest DC, here are the results during a deployment at ~2K rps. The GRPC version showed great improvements in CPU load and network consumption : 

### CPU usage 

As shown bellow, the GRPC version performed 40 % better than vanilla + shared blocking queries CPU-wise : 

Vanilla + SBQ : 
![image](https://user-images.githubusercontent.com/1712219/59280526-03370100-8c66-11e9-8bc1-10b0ba28b9dd.png)

GRPC :
![image](https://user-images.githubusercontent.com/1712219/59280561-12b64a00-8c66-11e9-8974-21bd4412e4fc.png)

Note that in addtion to overall CPU usage, system time is greatly reduced ( / 2.5 )

Comparison : 
![image](https://user-images.githubusercontent.com/1712219/59280644-32e60900-8c66-11e9-807f-1d752dfaba8e.png)

Load GRPC / Load Vanilla + SBQ : 
![image](https://user-images.githubusercontent.com/1712219/59280721-51e49b00-8c66-11e9-9272-501fd0057cdb.png)

## Network

The GRPC version performed 28% better in network consumption. 

Vanilla : 
![image](https://user-images.githubusercontent.com/1712219/59281168-16969c00-8c67-11e9-9a6d-1a1b051ada41.png)

GRPC:
![image](https://user-images.githubusercontent.com/1712219/59281050-e949ee00-8c66-11e9-98a6-27e9b85d0804.png)