@banks 2019-06-12
Thanks @Aestek and @pierresouchay, these results are really encouraging and great data for us!